+++ snapatac2-python/src/lib.rs
@@ -50,6 +50,7 @@ fn _snapatac2(_py: Python, m: &PyModule) -> PyResult<()> {
     m.add_function(wrap_pyfunction!(preprocessing::fragment_size_distribution, m)?)?;
 
     m.add_function(wrap_pyfunction!(export::export_fragments, m)?)?;
+    m.add_function(wrap_pyfunction!(export::export_bedgraph, m)?)?;
     m.add_function(wrap_pyfunction!(export::export_bigwig, m)?)?;
 
     m.add_function(wrap_pyfunction!(call_peaks::export_tags, m)?)?;